+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Several deprecated API entry points of TileDB Embedded are no longer used (#452, #453)
 
+* Support for DELETE queries has been added (requires TileDB Embedded 2.12.0 or later) (#455)
+
 ## Bug Fixes
 
 * Treatment of character columns with missing values has been corrected (#454)